Initial commit of tests
