+++ .codespell_words
@@ -1,0 +1,1 @@
+unexpect
