+++ functional_programming_tests/CMakeLists.txt
@@ -12,17 +12,17 @@ find_package(rclcpp REQUIRED)
 find_package(std_msgs REQUIRED)
 find_package(tl_expected REQUIRED)
 
-add_executable(main test/main.cpp)
-ament_target_dependencies(main rclcpp std_msgs example_srvs tl_expected)
-
 if(BUILD_TESTING)
   find_package(ament_cmake_gtest REQUIRED)
   find_package(ament_cmake_gmock REQUIRED)
-  ament_add_gtest(with_functional_programming test/with_functional_programming.cpp TIMEOUT 5)
-  ament_target_dependencies(with_functional_programming rclcpp std_msgs)
-
   ament_add_gtest(without_functional_programming test/without_functional_programming.cpp TIMEOUT 5)
   ament_target_dependencies(without_functional_programming rclcpp std_msgs example_srvs)
+
+  ament_add_gtest(with_functional_programming test/with_functional_programming.cpp TIMEOUT 5)
+  ament_target_dependencies(with_functional_programming rclcpp std_msgs example_srvs tl_expected)
+
+  ament_add_gtest(with_pure_functions test/with_pure_functions.cpp TIMEOUT 5)
+  ament_target_dependencies(with_pure_functions rclcpp std_msgs)
 
   ament_add_gmock(with_dependency_injection test/with_dependency_injection.cpp TIMEOUT 5)
   ament_target_dependencies(with_dependency_injection rclcpp std_msgs example_srvs)
